Tortoise-ORM FastAPI Integration/requirements.txt
fastapi
tortoise-orm
uvicorn